+++ demo/app/main-view-model.ts
@@ -39,10 +39,10 @@ export class HelloWorldModel extends Observable {
           preferredBarTintColor: '#453AA4',
           preferredControlTintColor: 'white',
           readerMode: false,
-          animated: false,
+          animated: true,
           modalPresentationStyle: 'overFullScreen',
           modalTransitionStyle: 'partialCurl',
-          modalEnabled: false,
+          modalEnabled: true,
           // Android Properties
           showTitle: true,
           toolbarColor: '#6200EE',