+++ src/InAppBrowser.ios.ts
@@ -100,7 +100,7 @@ const InAppBrowser = (<any>NSObject).extend({
       }
 
       const ctrl = UIApplication.sharedApplication.keyWindow.rootViewController;
-      
+
       if (options.modalEnabled) {
         safariVC.modalPresentationStyle = getPresentationStyle(options.modalPresentationStyle);
         if (options.animated) {
@@ -185,8 +185,27 @@ const InAppBrowser = (<any>NSObject).extend({
       this.close();
     }
   },
-  safariViewControllerDidFinish(controller: SFSafariViewController): void {
-    controller.dismissViewControllerAnimatedCompletion(this.animated, null);
+  dismissWithoutAnimation(controller: SFSafariViewController): void {
+    const animationKey = 'dismissInAppBrowser';
+    const ctrl = UIApplication.sharedApplication.keyWindow.rootViewController;
+    const transition = CATransition.animation();
+    transition.duration = 0.0;
+    transition.timingFunction = CAMediaTimingFunction.functionWithName(kCAMediaTimingFunctionEaseInEaseOut);
+    transition.type = kCATransitionFade;
+    transition.subtype = kCATransitionFromBottom;
+    controller.view.alpha = 0.05;
+    controller.view.frame = CGRectMake(0.0, 0.0, 0.5, 0.5);
+    ctrl.view.layer.addAnimationForKey(transition, animationKey);
+    ctrl.dismissViewControllerAnimatedCompletion(false, () => {
+      ctrl.view.layer.removeAnimationForKey(animationKey);
+    });
+  },
+  safariViewControllerDidFinish(
+    controller: SFSafariViewController
+  ): void {
+    if (!this.animated) {
+      this.dismissWithoutAnimation(controller);
+    }
     if (this.redirectResolve) {
       this.redirectResolve({
         type: 'cancel'
@@ -194,6 +213,7 @@ const InAppBrowser = (<any>NSObject).extend({
       this.flowDidFinish();
     }
   },
+
   flowDidFinish() {
     this.redirectResolve = null;
     this.redirectReject = null;